pyproject.toml
[project]
name = "airflow_project"
version = "0.1.0"
description = "Holistic ML Pipleline."
readme = "README.md"
requires-python = ">=3.8,<3.12"
dependencies = [
    "boto3==1.28.8",
    "pandas==2.2.3",
    "scikit-learn==1.5.2",
    "joblib==1.3.2",
    "mlflow==2.16.2",
    "dagster==1.8.11",
    "requests==2.32.3",
    "prometheus_client==0.21.0",
    "fastapi==0.115.2",
    "uvicorn==0.31.1",
    "apache-airflow[mlflow]",
    "influxdb==5.3.2",
    "docker==7.1.0",
    "kubernetes==31.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest"
]

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["data"]